refactor(mitm-addon): split stream usage tests
