+++ crates/runner/mitm-addon/tests/test_model_provider_sse_usage.py
@@ -1,0 +1,365 @@
+"""Tests for model-provider SSE usage reporting paths."""
+
+import json
+import uuid
+from collections.abc import Callable
+from pathlib import Path
+
+import pytest
+from mitmproxy import http
+from mitmproxy.flow import Error
+from mitmproxy.test import tutils
+
+import mitm_addon
+import usage
+from tests.flow_helpers import header_map, response_stream
+
+
+def _model_provider_sse_flow(
+    tmp_path: Path,
+    real_flow: Callable[..., http.HTTPFlow],
+    *,
+    host: str,
+    original_url: str,
+    firewall_name: str,
+    cli_agent_type: str | None = None,
+) -> http.HTTPFlow:
+    flow = real_flow(with_response=False, host=host)
+    flow.metadata["vm_run_id"] = "run-abc-123"
+    flow.metadata["vm_network_log_path"] = str(tmp_path / "network.jsonl")
+    flow.metadata["vm_proxy_log_path"] = str(tmp_path / "proxy.jsonl")
+    flow.metadata["firewall_action"] = "ALLOW"
+    flow.metadata["original_url"] = original_url
+    flow.metadata["firewall_name"] = firewall_name
+    flow.metadata["firewall_billable"] = True
+    flow.metadata["vm_sandbox_token"] = "tok-xyz"
+    if cli_agent_type is not None:
+        flow.metadata["cli_agent_type"] = cli_agent_type
+    flow.response = tutils.tresp(
+        status_code=200,
+        headers=header_map({"content-type": "text/event-stream"}),
+    )
+
+    mitm_addon.responseheaders(flow)
+    return flow
+
+
+def _openai_responses_sse_flow(
+    tmp_path: Path, real_flow: Callable[..., http.HTTPFlow]
+) -> http.HTTPFlow:
+    return _model_provider_sse_flow(
+        tmp_path,
+        real_flow,
+        host="api.openai.com",
+        original_url="https://api.openai.com/v1/responses",
+        firewall_name="model-provider:openai-api-key",
+        cli_agent_type="codex",
+    )
+
+
+def _model_sse_parse_warnings(flow: http.HTTPFlow) -> list[dict]:
+    proxy_log = Path(flow.metadata["vm_proxy_log_path"])
+    if not proxy_log.exists():
+        return []
+    return [
+        entry
+        for entry in (json.loads(line) for line in proxy_log.read_text().splitlines())
+        if entry.get("message") == "Model provider SSE usage extraction failed"
+    ]
+
+
+def _assert_single_model_sse_parse_warning(
+    flow: http.HTTPFlow,
+    *,
+    usage_protocol: str,
+    event: str,
+) -> None:
+    usage_warnings = _model_sse_parse_warnings(flow)
+    assert len(usage_warnings) == 1
+    warning = usage_warnings[0]
+    assert warning["level"] == "warn"
+    assert warning["type"] == "usage_event"
+    assert warning["usage_protocol"] == usage_protocol
+    assert warning["event"] == event
+    assert warning["error"]
+
+
+class TestModelProviderSseUsage:
+    """Tests for model-provider SSE usage reporting."""
+
+    @pytest.fixture(autouse=True)
+    def _sync_usage_delivery(self, sync_usage_executor, usage_webhook_api):
+        self._usage_webhook_api = usage_webhook_api
+
+    def _run_response(self, flow: http.HTTPFlow):
+        with self._usage_webhook_api() as webhook:
+            mitm_addon.response(flow)
+            usage.flush_usage_events(trigger="test")
+        return webhook
+
+    def _run_error(self, flow: http.HTTPFlow):
+        with self._usage_webhook_api() as webhook:
+            mitm_addon.error(flow)
+            usage.flush_usage_events(trigger="test")
+        return webhook
+
+    def test_full_pipeline_model_sse_finalizes_trailing_event(self, tmp_path, real_flow):
+        """response() must flush a trailing SSE usage event before reporting."""
+        flow = _openai_responses_sse_flow(tmp_path, real_flow)
+        response_stream(flow)(
+            b"event: response.completed\n"
+            b'data: {"response":{"model":"gpt-5.5",'
+            b'"usage":{"input_tokens":50,"output_tokens":20,'
+            b'"input_tokens_details":{"cached_tokens":10}}}}'
+        )
+
+        webhook = self._run_response(flow)
+
+        events = webhook.usage_events()
+        by_category = {event["category"]: event["quantity"] for event in events}
+        assert by_category == {
+            "tokens.input": 40,
+            "tokens.output": 20,
+            "tokens.cache_read": 10,
+        }
+
+    def test_full_pipeline_model_sse_reports_response_incomplete_usage(self, tmp_path, real_flow):
+        flow = _openai_responses_sse_flow(tmp_path, real_flow)
+        response_stream(flow)(
+            b"event: response.incomplete\n"
+            b'data: {"response":{"id":"resp_incomplete","model":"gpt-5.5",'
+            b'"usage":{"input_tokens":8000,"output_tokens":1024,'
+            b'"input_tokens_details":{"cached_tokens":2000}}}}\n\n'
+        )
+
+        webhook = self._run_response(flow)
+
+        events = webhook.usage_events()
+        by_category = {event["category"]: event["quantity"] for event in events}
+        assert by_category == {
+            "tokens.input": 6000,
+            "tokens.output": 1024,
+            "tokens.cache_read": 2000,
+        }
+        assert {event["provider"] for event in events} == {"gpt-5.5"}
+
+    def test_full_pipeline_anthropic_sse_logs_truncated_message_start(self, tmp_path, real_flow):
+        flow = _model_provider_sse_flow(
+            tmp_path,
+            real_flow,
+            host="api.anthropic.com",
+            original_url="https://api.anthropic.com/v1/messages",
+            firewall_name="model-provider:anthropic-api-key",
+        )
+        response_stream(flow)(
+            b'event: message_start\ndata: {"type":"message_start","message":{"id":"msg_1","mod'
+        )
+
+        webhook = self._run_response(flow)
+
+        assert webhook.request_count == 0
+        _assert_single_model_sse_parse_warning(
+            flow,
+            usage_protocol="anthropic_messages_sse",
+            event="message_start",
+        )
+
+    def test_full_pipeline_anthropic_sse_error_logs_truncated_message_start(
+        self, tmp_path, real_flow
+    ):
+        flow = _model_provider_sse_flow(
+            tmp_path,
+            real_flow,
+            host="api.anthropic.com",
+            original_url="https://api.anthropic.com/v1/messages",
+            firewall_name="model-provider:anthropic-api-key",
+        )
+        response_stream(flow)(
+            b'event: message_start\ndata: {"type":"message_start","message":{"id":"msg_1","mod'
+        )
+        flow.error = Error("connection reset by peer")
+
+        webhook = self._run_error(flow)
+
+        assert webhook.request_count == 0
+        _assert_single_model_sse_parse_warning(
+            flow,
+            usage_protocol="anthropic_messages_sse",
+            event="message_start",
+        )
+
+    def test_full_pipeline_anthropic_sse_logs_malformed_message_start(self, tmp_path, real_flow):
+        flow = _model_provider_sse_flow(
+            tmp_path,
+            real_flow,
+            host="api.anthropic.com",
+            original_url="https://api.anthropic.com/v1/messages",
+            firewall_name="model-provider:anthropic-api-key",
+        )
+        response_stream(flow)(b"event: message_start\ndata: {invalid json}\n\n")
+
+        webhook = self._run_response(flow)
+
+        assert webhook.request_count == 0
+        _assert_single_model_sse_parse_warning(
+            flow,
+            usage_protocol="anthropic_messages_sse",
+            event="message_start",
+        )
+
+    def test_full_pipeline_anthropic_sse_logs_truncated_message_delta_after_start(
+        self, tmp_path, real_flow
+    ):
+        flow = _model_provider_sse_flow(
+            tmp_path,
+            real_flow,
+            host="api.anthropic.com",
+            original_url="https://api.anthropic.com/v1/messages",
+            firewall_name="model-provider:anthropic-api-key",
+        )
+        response_stream(flow)(
+            b"event: message_start\n"
+            b'data: {"type":"message_start","message":{"id":"msg_1",'
+            b'"model":"claude-sonnet-4-6","usage":{"input_tokens":50}}}\n\n'
+            b"event: message_delta\n"
+            b'data: {"type":"message_delta","usage":{"output_tokens":'
+        )
+
+        webhook = self._run_response(flow)
+
+        events = webhook.usage_events()
+        by_category = {event["category"]: event["quantity"] for event in events}
+        assert by_category == {"tokens.input": 50}
+        assert {event["provider"] for event in events} == {"claude-sonnet-4-6"}
+        _assert_single_model_sse_parse_warning(
+            flow,
+            usage_protocol="anthropic_messages_sse",
+            event="message_delta",
+        )
+
+    def test_full_pipeline_openai_sse_logs_truncated_terminal_event(self, tmp_path, real_flow):
+        flow = _openai_responses_sse_flow(tmp_path, real_flow)
+        response_stream(flow)(
+            b"event: response.completed\n"
+            b'data: {"type":"response.completed","response":{"id":"resp_1","model":"gpt'
+        )
+
+        webhook = self._run_response(flow)
+
+        assert webhook.request_count == 0
+        _assert_single_model_sse_parse_warning(
+            flow,
+            usage_protocol="openai_responses_sse",
+            event="response.completed",
+        )
+
+    def test_full_pipeline_openai_sse_logs_truncated_late_event_name(self, tmp_path, real_flow):
+        flow = _openai_responses_sse_flow(tmp_path, real_flow)
+        response_stream(flow)(
+            b'data: {"type":"response.completed","response":{"id":"resp_1","model":"gpt\n'
+            b"event: response.completed\n\n"
+        )
+
+        webhook = self._run_response(flow)
+
+        assert webhook.request_count == 0
+        _assert_single_model_sse_parse_warning(
+            flow,
+            usage_protocol="openai_responses_sse",
+            event="response.completed",
+        )
+
+    def test_full_pipeline_eventless_incomplete_anthropic_usage_sse_warns(
+        self, tmp_path, real_flow
+    ):
+        flow = _model_provider_sse_flow(
+            tmp_path,
+            real_flow,
+            host="api.anthropic.com",
+            original_url="https://api.anthropic.com/v1/messages",
+            firewall_name="model-provider:anthropic-api-key",
+        )
+        response_stream(flow)(
+            b'data: {"type":"message_start","message":{"id":"msg_1","model":"claude'
+        )
+
+        webhook = self._run_response(flow)
+
+        assert webhook.request_count == 0
+        _assert_single_model_sse_parse_warning(
+            flow,
+            usage_protocol="anthropic_messages_sse",
+            event="message_start",
+        )
+
+    def test_full_pipeline_anthropic_non_usage_incomplete_sse_does_not_warn(
+        self, tmp_path, real_flow
+    ):
+        flow = _model_provider_sse_flow(
+            tmp_path,
+            real_flow,
+            host="api.anthropic.com",
+            original_url="https://api.anthropic.com/v1/messages",
+            firewall_name="model-provider:anthropic-api-key",
+        )
+        response_stream(flow)(
+            b"event: content_block_delta\n"
+            b'data: {"type":"content_block_delta","delta":{"text":"hello'
+        )
+
+        webhook = self._run_response(flow)
+
+        assert webhook.request_count == 0
+        assert _model_sse_parse_warnings(flow) == []
+
+    def test_full_pipeline_openai_eventless_incomplete_sse_does_not_warn(self, tmp_path, real_flow):
+        flow = _openai_responses_sse_flow(tmp_path, real_flow)
+        response_stream(flow)(
+            b'data: {"type":"response.completed","response":{"id":"resp_1","model":"gpt'
+        )
+
+        webhook = self._run_response(flow)
+
+        assert webhook.request_count == 0
+        assert _model_sse_parse_warnings(flow) == []
+
+    def test_full_pipeline_openai_non_terminal_incomplete_sse_does_not_warn(
+        self, tmp_path, real_flow
+    ):
+        flow = _openai_responses_sse_flow(tmp_path, real_flow)
+        response_stream(flow)(
+            b"event: response.in_progress\n"
+            b'data: {"type":"response.in_progress","response":{"id":"resp_1","model":"gpt'
+        )
+
+        webhook = self._run_response(flow)
+
+        assert webhook.request_count == 0
+        assert _model_sse_parse_warnings(flow) == []
+
+    def test_full_pipeline_model_sse_zero_event_preserves_billed_usage_and_id(
+        self, tmp_path, real_flow
+    ):
+        flow = _openai_responses_sse_flow(tmp_path, real_flow)
+        response_stream(flow)(
+            b"event: response.completed\n"
+            b'data: {"response":{"id":"resp_sse_1","model":"gpt-5.5",'
+            b'"usage":{"input_tokens":100,"output_tokens":40}}}\n\n'
+            b"event: response.failed\n"
+            b'data: {"response":{"id":"resp_sse_empty","model":"gpt-5.4",'
+            b'"usage":{"input_tokens":0,"output_tokens":0}}}\n\n'
+        )
+
+        webhook = self._run_response(flow)
+
+        events = webhook.usage_events()
+        by_category = {event["category"]: event["quantity"] for event in events}
+        idempotency_by_category = {event["category"]: event["idempotencyKey"] for event in events}
+        assert by_category == {
+            "tokens.input": 100,
+            "tokens.output": 40,
+        }
+        assert set(idempotency_by_category) == {"tokens.input", "tokens.output"}
+        for key in idempotency_by_category.values():
+            uuid.UUID(key)
+        assert {event["provider"] for event in events} == {"gpt-5.5"}
